Github workflow to build and publish a release.
This workflow will just build the project without creating a release
when a commit is pushed to master or a pull request is created.

It also publishes a release if manually triggered by a maintainer of the
project on a tag reference.

The tarball that's created contains an SPDK installation + the vhost_ubi
binary from the bdev_ubi project.

Github's actions/create-release@v1 [1] isn't maintained anymore and
fails with an error if used, so we used ncipollo/release-action to
publish releases.

This is based on Cloud-hypervisor's EDK2 fork's release.yaml [3], except
that we fix couple of things that doesn't work anymore:

1. Don't use actions/create-release@v1 action
2. `github.event.ref_type` => `github.ref_type`

[1] https://github.com/actions/create-release
[2] https://github.com/ncipollo/release-action
[3]
https://raw.githubusercontent.com/cloud-hypervisor/edk2/92c79b2901c24de6a0859b8ad4f0b1a2bfa79503/.github/workflows/release.yaml