Add valgrind tests.
Valgrind forces multi-threaded apps to run on a single-thread. Our
previous tests required to be run on multiple cores. This PR adds a test
that runs on a single core using continuation style programming so we
can run valgrind on top of it. This test exercises different code paths
to make sure their memory accesses are sane.

To run this, do:
```
SPDK_PATH=/path/to/spdk make valgrind
```