Use SPDK v23.09 instead of SPDK v24.01 in the build pipeline
Previously I had merged the change to use v24.01 because it passed
bdev_ubi regression tests and some local manual tests.

But when I tried it along with Ubicloud recently, it failed with:

```
spdk_rpc.rb:113:in `call': Input/output error (SpdkRpcError)
```

So, we change the SPDK version for releases back to SPDK v23.09 to
unblock releasing binaries for Ubuntu 24.04. The issues with newer
versions of SPDK will be fixed in a separate PR.

Note that we never tried SPDK v24.01 in production, so this isn't a
downgrade of what we are already running.