Instrument SPDK's iobuff init & free with debug info.
SPDK gets items from large and small pools when an IO channel is
created. We had crashes in production related to this, so logging the
ring buffer sizes will help with debugging. These logs are created only
a few times during a VM's lifetime, so they won't cause log spam.