Fix error handling & make errors testable.
When reading metadata, we open an IO channel to the base bdev. This can
return NULL, and not checking for NULL caused a crash in production.
This change adds a check to prevent the crash and includes similar
fixes.

It also adds functions to make errors in the create/delete paths
testable.